Support for stopping based on SongEnd or ResultsView
