+++ OBSControl/HarmonyPatches/HarmonyManager.cs
@@ -22,6 +22,7 @@ namespace OBSControl.HarmonyPatches
         private static readonly BindingFlags allBindingFlags = BindingFlags.Static | BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic;
         private static HarmonyPatchInfo? LevelDelayPatch;
         private static HarmonyPatchInfo? ReadyToStartPatch;
+        private static HarmonyPatchInfo? LevelDidFinishPatch;
         internal readonly static HashSet<HarmonyPatchInfo> AppliedPatches = new HashSet<HarmonyPatchInfo>();
 
         public static bool ApplyPatch(HarmonyPatchInfo patchInfo)
@@ -57,7 +58,7 @@ namespace OBSControl.HarmonyPatches
 
         public static void ApplyDefaultPatches()
         {
-
+            GetLevelDidFinishPatch().ApplyPatch();
         }
 
         public static void UnpatchAll()
@@ -87,6 +88,16 @@ namespace OBSControl.HarmonyPatches
                 ReadyToStartPatch = new HarmonyPatchInfo(Harmony, original, null, postFix);
             }
             return ReadyToStartPatch;
+        }
+        public static HarmonyPatchInfo GetLevelDidFinishPatch()
+        {
+            if (LevelDidFinishPatch == null)
+            {
+                MethodInfo original = typeof(LevelSelectionFlowCoordinator).GetMethod("HandleStandardLevelDidFinish", allBindingFlags);
+                HarmonyMethod postFix = new HarmonyMethod(typeof(HandleStandardLevelDidFinishPatch).GetMethod("Postfix", allBindingFlags));
+                LevelDidFinishPatch = new HarmonyPatchInfo(Harmony, original, null, postFix);
+            }
+            return LevelDidFinishPatch;
         }
     }
 }
