+++ OBSControl/OBSComponents/RecordingController.cs
@@ -82,14 +82,7 @@ namespace OBSControl.OBSComponents
         {
             get
             {
-                return StopRecordAction switch
-                {
-                    RecordActionType.Auto => RecordStopOption.SceneSequence,
-                    RecordActionType.NoAction => RecordStopOption.None,
-                    RecordActionType.Immediate => _recordStopOption,
-                    RecordActionType.None => _recordStopOption,
-                    _ => _recordStopOption
-                };
+                return AutoStop ? _recordStopOption : RecordStopOption.None;
             }
             set
             {
@@ -99,23 +92,18 @@ namespace OBSControl.OBSComponents
         /// <summary>
         /// When should record stop be triggered, if at all.
         /// </summary>
-        public RecordActionType StopRecordAction
+        public bool AutoStop
         {
             get
             {
-                return (RecordStartSource, AutoStopOnManual, DelayedStopEnabled) switch
+                return (RecordStartSource, AutoStopOnManual) switch
                 {
-                    (RecordActionSourceType.Manual, true, true) => RecordActionType.Delayed,
-                    (RecordActionSourceType.Manual, true, false) => RecordActionType.Immediate,
-                    (RecordActionSourceType.Manual, false, _) => RecordActionType.NoAction,
-                    (RecordActionSourceType.ManualOBS, true, true) => RecordActionType.Delayed,
-                    (RecordActionSourceType.ManualOBS, true, false) => RecordActionType.Immediate,
-                    (RecordActionSourceType.ManualOBS, false, _) => RecordActionType.NoAction,
-                    (RecordActionSourceType.Auto, _, _) => RecordActionType.Auto,
-                    (RecordActionSourceType.None, true, true) => RecordActionType.Delayed,
-                    (RecordActionSourceType.None, true, false) => RecordActionType.Immediate,
-                    (RecordActionSourceType.None, false, _) => RecordActionType.NoAction,
-                    _ => RecordActionType.None
+                    (RecordActionSourceType.Manual, true) => true,
+                    (RecordActionSourceType.Manual, false) => false,
+                    (RecordActionSourceType.ManualOBS, true) => true,
+                    (RecordActionSourceType.ManualOBS, false) => false,
+                    (RecordActionSourceType.Auto, _) => true,
+                    _ => false
 
                 };
             }
@@ -425,7 +413,6 @@ namespace OBSControl.OBSComponents
                 };
                 LastLevelData = recordingData;
             }
-            BS_Utils.Plugin.LevelDidFinishEvent += OnLevelFinished;
         }
 
         public async Task<RecordingSettings> SetupRecording(string? fileFormat, string? outputDirectory, CancellationToken cancellationToken)
@@ -525,124 +512,10 @@ namespace OBSControl.OBSComponents
         }
 
 
-        private async void OnLevelFinished(StandardLevelScenesTransitionSetupDataSO levelScenesTransitionSetupDataSO, LevelCompletionResults levelCompletionResults)
-        {
-            bool multipleLevelData = LastLevelData?.LevelResults != null || (LastLevelData?.MultipleLastLevels ?? false) == true;
-            try
-            {
-                PlayerLevelStatsData? stats = null;
-                IBeatmapLevel? levelInfo = GameStatus.LevelInfo;
-                IDifficultyBeatmap? difficultyBeatmap = GameStatus.DifficultyBeatmap;
-                PlayerDataModel? playerData = OBSController.instance?.PlayerData;
-                if (playerData != null && levelInfo != null && difficultyBeatmap != null)
-                {
-                    stats = playerData.playerData.GetPlayerLevelStatsData(
-                        levelInfo.levelID, difficultyBeatmap.difficulty, difficultyBeatmap.parentDifficultyBeatmapSet.beatmapCharacteristic);
-                }
-
-                LevelCompletionResultsWrapper levelResults = new LevelCompletionResultsWrapper(levelCompletionResults, stats?.playCount ?? 0, GameStatus.MaxModifiedScore);
-                RecordingData? recordingData = LastLevelData;
-                if (recordingData == null)
-                {
-                    recordingData = new RecordingData(new BeatmapLevelWrapper(difficultyBeatmap), levelResults, stats)
-                    {
-                        MultipleLastLevels = multipleLevelData
-                    };
-                    LastLevelData = recordingData;
-                }
-                else
-                {
-                    if (recordingData.LevelData == null)
-                    {
-                        recordingData.LevelData = new BeatmapLevelWrapper(difficultyBeatmap);
-                    }
-                    else if (difficultyBeatmap != null && recordingData.LevelData.DifficultyBeatmap != difficultyBeatmap)
-                    {
-                        Logger.log?.Debug($"Existing beatmap data doesn't match level completion beatmap data: '{recordingData.LevelData.SongName}' != '{difficultyBeatmap?.level.songName}'");
-                        recordingData.LevelData = new BeatmapLevelWrapper(difficultyBeatmap);
-                    }
-                    recordingData.LevelResults = levelResults;
-                    recordingData.PlayerLevelStats = stats;
-                    recordingData.MultipleLastLevels = multipleLevelData;
-                }
-
-            }
-#pragma warning disable CA1031 // Do not catch general exception types
-            catch (Exception ex)
-            {
-                Logger.log?.Error($"Error generating new file name: {ex}");
-                Logger.log?.Debug(ex);
-            }
-#pragma warning restore CA1031 // Do not catch general exception types
-            if (StopRecordAction == RecordActionType.Immediate)
-                StopRecordingTask = TryStopRecordingAsync();
-            else if (StopRecordAction == RecordActionType.Delayed)
-            {
-                TimeSpan stopDelay = TimeSpan.FromSeconds(Plugin.config?.RecordingStopDelay ?? 0);
-                if (stopDelay > TimeSpan.Zero)
-                    await Task.Delay(stopDelay);
-                await TryStopRecordingAsync();
-            }
-        }
-
-        #region OBS Event Handlers
-
-
-
-        private void OnObsRecordingStateChanged(object sender, OutputState type)
-        {
-            Logger.log?.Info($"Recording State Changed: {type}");
-            OutputState = type;
-            LastRecordingStateUpdate = DateTime.UtcNow;
-            switch (type)
-            {
-                case OutputState.Starting:
-                    recordingCurrentLevel = true;
-                    break;
-                case OutputState.Started:
-                    RecordStartTime = DateTime.UtcNow;
-                    recordingCurrentLevel = true;
-                    if (RecordStartSource == RecordActionSourceType.None)
-                    {
-                        RecordStartSource = RecordActionSourceType.ManualOBS;
-                        RecordStartOption = RecordStartOption.None;
-                    }
-                    Task.Run(() => Obs.GetConnectedObs()?.SetFilenameFormatting(DefaultFileFormat));
-                    break;
-                case OutputState.Stopping:
-                    recordingCurrentLevel = false;
-                    break;
-                case OutputState.Stopped:
-                    recordingCurrentLevel = false;
-                    RecordStartTime = DateTime.MaxValue;
-                    RecordingData? lastLevelData = LastLevelData;
-                    string? renameOverride = RenameStringOverride;
-                    RenameStringOverride = null;
-                    LastLevelData = null;
-                    RecordStartSource = RecordActionSourceType.None;
-                    RecordStartOption = RecordStartOption.None;
-                    string? renameString = renameOverride ??
-                        lastLevelData?.GetFilenameString(Plugin.config.RecordingFileFormat, Plugin.config.InvalidCharacterSubstitute, Plugin.config.ReplaceSpacesWith);
-                    if (renameString != null)
-                        RenameLastRecording(renameString);
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        #endregion
 
         #region Setup/Teardown
 
 
-        private void OnSceneStageChanged(object sender, SceneStageChangedEventArgs e)
-        {
-#if DEBUG
-            Logger.log?.Debug($"RecordingController: OnSceneStageChanged - {e.SceneStage}.");
-#endif
-            e.AddCallback(SceneSequenceCallback);
-        }
         private void OnOBSComponentChanged(object sender, OBSComponentChangedEventArgs e)
         {
             if (e.AddedComponent is SceneController addedSceneController)
@@ -670,7 +543,9 @@ namespace OBSControl.OBSComponents
             obs.RecordingStateChanged += OnObsRecordingStateChanged;
             obs.OBSComponentChanged += OnOBSComponentChanged;
             StartLevelPatch.LevelStarting += OnLevelStarting;
+            HandleStandardLevelDidFinishPatch.LevelDidFinish += OnLevelDidFinish;
             BSEvents.gameSceneActive += OnGameSceneActive;
+            BS_Utils.Plugin.LevelDidFinishEvent += OnLevelFinished;
         }
 
 
@@ -681,80 +556,21 @@ namespace OBSControl.OBSComponents
             obs.RecordingStateChanged -= OnObsRecordingStateChanged;
             obs.OBSComponentChanged -= OnOBSComponentChanged;
             StartLevelPatch.LevelStarting -= OnLevelStarting;
+            HandleStandardLevelDidFinishPatch.LevelDidFinish -= OnLevelDidFinish;
             BSEvents.gameSceneActive -= OnGameSceneActive;
-        }
-
-        private void OnLevelStarting(object sender, LevelStartingEventArgs e)
-        {
-            RecordStartOption recordStartOption = RecordStartOption;
-            Logger.log?.Debug($"RecordingController OnLevelStarting. StartOption is {recordStartOption}");
-            switch (recordStartOption)
-            {
-                case RecordStartOption.None:
-                    break;
-                case RecordStartOption.SceneSequence:
-                    break;
-                case RecordStartOption.SongStart:
-                    break;
-                case RecordStartOption.LevelStartDelay:
-                    e.SetResponse(LevelStartingSourceName, (int)(RecordingStartDelay * 1000));
-                    break;
-                case RecordStartOption.Immediate:
-                    break;
-                default:
-                    break;
-            }
-        }
-
-
-        private async void OnGameSceneActive()
-        {
-            Logger.log?.Debug($"RecordingController OnGameSceneActive.");
-            StartCoroutine(GameStatusSetup());
-            if (RecordStartOption == RecordStartOption.SongStart)
-            {
-                await TryStartRecordingAsync(RecordActionSourceType.Auto, RecordStartOption.SongStart);
-            }
-            //var timeControllers = Resources.FindObjectsOfTypeAll<AudioTimeSyncController>();
-            //var timeController = timeControllers.First();
-            //float oldTimeScale = SyncControllerTimeScale(ref timeController);
-            //Logger.log?.Debug($"Setting TimeScale to '0', old TimeScale is '{oldTimeScale}'");
-            //SyncControllerTimeScale(ref timeController) = 0;
-            //await Task.Delay(1000);
-            //Logger.log?.Debug($"Setting TimeScale back to '{oldTimeScale}'");
-            //SyncControllerTimeScale(ref timeController) = oldTimeScale;
-            //var songControllers = Resources.FindObjectsOfTypeAll<SongController>();
-            //if (songControllers.Length > 1)
-            //    Logger.log?.Error($"{songControllers.Length} SongControllers exist.");
-            //var songController = songControllers?.FirstOrDefault();
-            //var pauseController = Resources.FindObjectsOfTypeAll<GamePause>().FirstOrDefault();
-            //if (songController != null)
-            //{
-            //    await Task.Delay(500);
-            //    songController.StopSong();
-            //    Logger.log?.Debug($"RecordingController song stopped? delaying by 10s.");
-            //    await Task.Delay(10000);
-            //    songController.StartSong();
-            //}
-            //else
-            //    Logger.log?.Debug("timeController is null.");
+            BS_Utils.Plugin.LevelDidFinishEvent -= OnLevelFinished;
         }
 
         private async Task SceneSequenceCallback(SceneStage sceneStage)
         {
             Logger.log?.Debug($"RecordingController: SceneStage - {sceneStage}.");
-            if (sceneStage == SceneStage.IntroStarted)
+            if (sceneStage == SceneStage.IntroStarted && RecordStartOption == RecordStartOption.SceneSequence)
             {
                 await TryStartRecordingAsync(RecordActionSourceType.Auto, RecordStartOption.SceneSequence);
             }
-            else if (sceneStage == SceneStage.Game)
-                StartCoroutine(GameStatusSetup());
             else if (sceneStage == SceneStage.OutroFinished)
             {
-                //TimeSpan recordStopDelay = TimeSpan.FromSeconds(Plugin.config?.RecordingStopDelay ?? 0);
-                //if (recordStopDelay > TimeSpan.Zero)
-                //    await Task.Delay(recordStopDelay); // TODO: this also forces scene sequence to wait, do I even need this with the scene settings?
-                if (StopRecordAction == RecordActionType.Auto)
+                if (RecordStopOption == RecordStopOption.SceneSequence)
                     await TryStopRecordingAsync();
             }
         }
