+++ OBSControl/OBSControl.csproj
@@ -139,7 +139,9 @@
     <Compile Include="HarmonyPatches\GameSongController_ReadyToStart.cs" />
     <Compile Include="HarmonyPatches\HarmonyPatchInfo.cs" />
     <Compile Include="HarmonyPatches\LevelSelectionFlowCoordinator_StartLevel.cs" />
+    <Compile Include="HarmonyPatches\LevelSelectionFlowCoordinator_HandleStandardLevelDidFinishPatch.cs" />
     <Compile Include="OBSComponents\OBSComponent.cs" />
+    <Compile Include="OBSComponents\RecordingController_EventHandlers.cs" />
     <Compile Include="OBSComponents\RecordingController_Types.cs" />
     <Compile Include="OBSComponents\SceneController.cs" />
     <Compile Include="OBSComponents\StreamingController.cs" />
